pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Mono-user self-hosted second brain unifying news, notes, flashcards and tasks behind an LLM chatbot."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "httpx>=0.28",
    "bcrypt>=4.2",
    "itsdangerous>=2.2",
    "python-multipart>=0.0.18",
    "python-frontmatter>=1.1",
    "python-ulid>=3.0",
    "click>=8.1",
]

[project.scripts]
second-brain = "app.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "mypy>=1.13",
    "respx>=0.21",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
pythonpath = ["backend"]
